initial commit DONE: - test base init - some tests. need more TODO: - custom methods - tests - change db and test

+++ src/main/java/org/klozevitz/kameleoon_test/model/dao/daoDB/IDaoDB.java
@@ -1,6 +1,6 @@
 package org.klozevitz.kameleoon_test.model.dao.daoDB;
 
 public interface IDaoDB<E> {
-    E findById(Long id);
+    E findById(int id);
     E save(E e);
 }
